+++ atklip/graph_objects/chart_component/indicators/basic_indicators/optimation_ma.py
@@ -94,7 +94,7 @@ class BasicMA(GraphicsObject):
         
         df:pd.DataFrame = self.has["inputs"]["source"].get_df()
         self._INDICATOR = ta.ma(f"{self.has["inputs"]["ma_type"].name}".lower(), df[f"{self.has["inputs"]["type"]}"],length=self.has["inputs"]["period"])
-        self._INDICATOR = self._INDICATOR.astype('float64')
+        self._INDICATOR = self._INDICATOR.astype('float32')
         _data = self._INDICATOR.to_numpy()
         _index = df["index"].to_numpy()
 
@@ -197,7 +197,7 @@ class BasicMA(GraphicsObject):
         df:pd.DataFrame = self.has["inputs"]["source"].get_df()
         self._INDICATOR = ta.ma(f"{self.has["inputs"]["ma_type"].name}".lower(), df[f"{self.has["inputs"]["type"]}"],length=self.has["inputs"]["period"])
         
-        self._INDICATOR = self._INDICATOR.astype('float64')
+        self._INDICATOR = self._INDICATOR.astype('float32')
         
         _data = self._INDICATOR.to_numpy()
         _index = df["index"].to_numpy()
@@ -353,7 +353,7 @@ class BasicMA(GraphicsObject):
     def update_data(self,last_candle:List[OHLCV],setdata):
         df:pd.DataFrame = self.has["inputs"]["source"].get_df()
         self._INDICATOR = ta.ma(f"{self.has["inputs"]["ma_type"].name}".lower(), df[f"{self.has["inputs"]["type"]}"],length=self.has["inputs"]["period"])
-        self._INDICATOR = self._INDICATOR.astype('float64')
+        self._INDICATOR = self._INDICATOR.astype('float32')
         _data = self._INDICATOR.to_numpy()
         _index = df["index"].to_numpy()
         setdata.emit((_index,_data))
